requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
Pillow>=9.0.0
requests>=2.25.0
